requirements.txt
paho-mqtt
slacker
